Fixed a bug where the first point being tracked could be null. Changed CommonUtils to an object. Changed Add run type target distance to Float

+++ app/src/main/java/com/example/coursework/ui/dashboard/DashboardScreen.kt
@@ -82,7 +82,7 @@ fun DashboardScreen(
     onStartRunSelected: (RunType) -> Unit,
     onFilterSelected: (String) -> Unit,
     onRunTypeSelected: (String) -> Unit,
-    onAddNewRunType: (String, Int) -> Unit
+    onAddNewRunType: (String, Float) -> Unit
 ) {
     var showAddRunTypeSheet by remember { mutableStateOf(false) }
     var showRunTypePicker by remember { mutableStateOf(false) }
