+++ app/src/main/java/com/example/coursework/ui/navigation/AppNavHost.kt
@@ -52,8 +52,8 @@ fun AppNavHost() {
                 },
                 onFilterSelected = {},// Placeholder until filtering is implemented
                 onRunTypeSelected = { vm.onRunTypeSelected(it) },
-                onAddNewRunType = { name, distance -> 
-                    addRunTypeVm.addRunType(name, distance.toFloat())
+                onAddNewRunType = { name, distance ->
+                    addRunTypeVm.addRunType(name, distance)
                 }
             )
         }
